# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MVP multi-tenant event platform. Auth owns user identity; org membership and permissions live in organization/iam.",
        "title": "Rallya API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/orgs/{id}/events/{eventId}/kit-collections": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter by kitId, status (PENDING|COLLECTED|VOIDED), or attendeeId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kits"
                ],
                "summary": "List handouts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event UUID or slug",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kit UUID",
                        "name": "kitId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Collection status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Attendee UUID",
                        "name": "attendeeId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/kitdto.CollectionListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/orgs/{id}/events/{eventId}/kit-collections/{collectionId}/collect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "PENDING -\u003e COLLECTED with staff stamps. Anything else is 422.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kits"
                ],
                "summary": "Mark a reserved handout collected",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event UUID or slug",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Collection UUID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/kitdto.CollectionResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/orgs/{id}/events/{eventId}/kit-collections/{collectionId}/void": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "PENDING or COLLECTED -\u003e VOIDED (frees re-issue). Anything else is 422.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kits"
                ],
                "summary": "Void a handout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event UUID or slug",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Collection UUID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/kitdto.CollectionResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/orgs/{id}/events/{eventId}/kits": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Every kit type for one event with live pending/collected/voided/remaining tallies.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kits"
                ],
                "summary": "List kit types",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event UUID or slug",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/kitdto.KitResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Named kit (merch/welcome pack) with a total quantity for one event.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kits"
                ],
                "summary": "Define a kit type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event UUID or slug",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Kit definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kitdto.CreateKitRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/kitdto.KitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    }
                }
            }
        },
...
        "/orgs/{id}/events/{eventId}/kits/{kitId}/collect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Attendee must be CHECKED_IN (422 otherwise). Collects immediately unless reserve=true holds a PENDING unit. IdempotencyKey makes tablet retries safe.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kits"
                ],
                "summary": "Hand a kit to an attendee",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event UUID or slug",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kit UUID",
                        "name": "kitId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Attendee handout",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kitdto.CollectRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/kitdto.CollectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/kitdto.ErrorAlias"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "kitdto.CollectRequest": {
            "type": "object",
            "properties": {
                "attendeeId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "idempotencyKey": {
                    "type": "string",
                    "example": "kit-req-001"
                },
                "reserve": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "kitdto.CollectionListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/kitdto.CollectionResponse"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "perPage": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 40
                }
            }
        },
        "kitdto.CollectionResponse": {
            "type": "object",
            "properties": {
                "attendeeId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "collectedAt": {
                    "type": "string",
                    "example": "2026-09-20T10:00:00Z"
                },
                "collectedBy": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-09-20T10:00:00Z"
                },
                "eventId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "kitId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "kitName": {
                    "type": "string",
                    "example": "VIP pack"
                },
                "status": {
                    "type": "string",
                    "example": "COLLECTED"
                }
            }
        },
        "kitdto.CreateKitRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Lanyard, shirt, stickers"
                },
                "name": {
                    "type": "string",
                    "example": "VIP pack"
                },
                "quantityTotal": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "kitdto.ErrorAlias": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "kit not found"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003caccessToken\u003e from POST /auth/login (or /auth/refresh).\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}